chore: major refactoring of ExpectationManager
