+++ apps/package-manager/packages/generic/src/generateExpectations/nrk/lib.ts
@@ -1,2 +1,10 @@
+/**
+ * Magic PackageContainer ID.
+ * When a PackageContainer with this ID is present, it'll be used as for the "smartbull" feature.
+ */
 export const SMARTBULL_STORAGE_ID = 'source-smartbull'
+/**
+ * Magic Temporary storage ID.
+ * When a PackageContainer with this ID is present, it'll be used as for the "smartbull" feature.
+ */
 export const TEMPORARY_STORAGE_ID = 'temporary-storage'
